crosshair fix

+++ source/nucledian/engine/graphics/shaders/shaders.h
@@ -117,8 +117,8 @@ namespace nc
       inline constexpr Uniform<2, vec2> TEXTURE_POS;
       inline constexpr Uniform<3, vec2> TEXTURE_SIZE;
       inline constexpr Uniform<4, s32> CHARACTER;
-      inline constexpr Uniform<5, f32> WIDTH;
-      inline constexpr Uniform<6, f32> HEIGHT;
+      inline constexpr Uniform<5, s32> WIDTH;
+      inline constexpr Uniform<6, s32> HEIGHT;
       
     }
